Add Cyclic Group Representation (CGR) class (#182)
* Add BaseMCRTensor

Add a new base class to support future implementation of variations of
MCR proposals.

* Fix bug in BaseMCRTensor:__torch_function__()

There is a bug when args is a collection of collections instead of a
plain tuple. In this case, the old args parsing was unable to search for
block_size in nested structures containing BaseMCRTensor.

For instance, let `a` and `b` be BaseMCRTensor variables. Calling
`torch.stack((a, b))` results in error, as the `args` received in
`__torch_function__()` is a nested tuple.

* Add CGRTensor

Add a new VSA class named Cyclic Group Representation (CGR). This class
is similar to MCR, but differs in bundling.

* Add CGRTensor to functional embeddings

Allow its usage in level and circular embeddings as done with MCRTensor.

* Fix CGR bundle

Ensure both inputs are in the same shape.

* Add CGR into tests

The CGR should behave almost the same as MCR, but diverges in bundling.
Implement a custom bundling test for it.

* Use Tensor.dim() in CGRTensor

Shorter syntax and more readable.

* Remove asserts in CGR/MCR classes

No need to check block size in CGR/MCR functions as the
__torch_function__() in BaseMCRTensor already checks it.

* Improve exception in BaseMCRTensor if multiple block_sizes

- Raise ValueError instead of RuntimeError

* Add CGRTensor to docs

Include CGRTensor in the README and in the built docs.

* Add BaseMCRTensor to docs

* Fix typo in README.md

* Improve documentation for MCRTensor and CGRTensor

Replicate function signatures in inherited classes (MCRTensor and
CGRTensor) from their parent class (BaseMCRTensor) to let sphinx
print the documentation of each parent function.